+++ src/main/java/kuit/subway/controller/AuthController.java
@@ -10,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import org.apache.coyote.Response;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.io.IOException;
 import java.net.URI;
@@ -24,12 +25,6 @@ public class AuthController {
     public ResponseEntity<TokenResponse> login(@RequestBody TokenRequest request) {
         TokenResponse tokenResponse = authService.createToken(request);
         return ResponseEntity.created(URI.create("/token")).body(tokenResponse);
-    }
-
-    @GetMapping("/github/callback")
-    public ResponseEntity<String> getGithubToken(@RequestParam String code) throws IOException {
-        String token = authService.getGithubToken(code);
-        return ResponseEntity.ok().body(token);
     }
 
 }
